+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### Bug fix
+
+ * Refactor remove interface Scope
+
 ## v0.0.1 (2025-03-18)
 
 ### Features